+++ src/test/java/store/lastdance/controller/user/UserV2ControllerTest.java
@@ -19,7 +19,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import store.lastdance.converter.UserConverter;
+import store.lastdance.converter.user.UserConverter;
 import store.lastdance.domain.user.User;
 import store.lastdance.dto.user.UserResponseDTO;
 import store.lastdance.dto.user.UserUpdateRequestDTO;
